feat: creat module profile

+++ src/routes/role/role.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common'
+import { BadRequestException, Injectable } from '@nestjs/common'
 import { RoleRepo } from './role.repo'
 import { RoleQueryType } from 'src/routes/role/role.model'
 import { CreateRoleBodyType, GetRoleResType, UpdateRoleBodyType } from 'src/routes/role/role.model'
-import { RoleAlreadyExistsException, RoleNotFoundException } from 'src/routes/role/role.error'
+import { ProhibitedRoleException, RoleAlreadyExistsException, RoleNotFoundException } from 'src/routes/role/role.error'
 import { isNotFoundPrismaError, isUniqueContraintPrismaError } from 'src/shared/helpers'
+import { ROLE_NAME } from 'src/shared/constants/roleName.constant'
 @Injectable()
 export class RoleService {
   constructor(private readonly roleRepo: RoleRepo) {}
@@ -38,8 +39,17 @@ export class RoleService {
 
   async update({ id, data, updatedById }: { id: number; data: UpdateRoleBodyType; updatedById: number }) {
     try {
-      const role = await this.roleRepo.update({ id, data, updatedById })
-      return role
+
+      const role = await this.roleRepo.getById(id)
+      if (!role) {
+        throw RoleNotFoundException
+      }
+      
+      if (role.name === ROLE_NAME.ADMIN) {
+        throw ProhibitedRoleException
+      }
+      const updatedRole = await this.roleRepo.update({ id, data, updatedById })
+      return updatedRole
     } catch (error) {
       if (isNotFoundPrismaError(error)) {
         throw RoleNotFoundException
@@ -47,12 +57,23 @@ export class RoleService {
       if (isUniqueContraintPrismaError(error)) {
         throw RoleAlreadyExistsException
       }
+      if (error instanceof Error) {
+        throw new BadRequestException(error.message)
+      }
       throw error
     }
   }
 
   async delete(id: number) {
     try {
+      const role = await this.roleRepo.getById(id)
+      if (!role) {
+        throw RoleNotFoundException
+      }
+      const baseRoles: string[] = [ROLE_NAME.ADMIN, ROLE_NAME.CLIENT, ROLE_NAME.SELLER]
+      if (baseRoles.includes(role.name)) {
+        throw ProhibitedRoleException
+      }
       await this.roleRepo.delete(id)
       return {
         message: 'Role deleted successfully',
